도메인 수정

+++ src/main/java/capstone3d/Server/api/FurnitureListApiController.java
@@ -43,10 +43,10 @@ public class FurnitureListApiController {
 
         private String furniture_img_url;
 
-        private int x;
+        private double x;
 
-        private int y;
+        private double y;
 
-        private int z;
+        private double z;
     }
 }
